gate authorization done, export to pdf in progress

+++ app/Http/Controllers/InformationController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Gate;
 
 class InformationController extends Controller
 {
@@ -15,15 +16,18 @@ class InformationController extends Controller
      */
     public function index(Request $request)
     {
-        $informations = \App\Information::paginate(10);
-        $no = 1;
+        if (Gate::allows('manage-informations')) {
+            $informations = \App\Information::paginate(10);
+            $no = 1;
 
-        $filterKeyword = $request->get('keyword');
-        if ($filterKeyword) {
-            $informations = \App\Information::where('title', 'LIKE', "%$filterKeyword%")->paginate(10);
+            $filterKeyword = $request->get('keyword');
+            if ($filterKeyword) {
+                $informations = \App\Information::where('title', 'LIKE', "%$filterKeyword%")->paginate(10);
+            }
+
+            return view('informations.index', ['informations' => $informations, 'nomor' => $no]);
         }
-
-        return view('informations.index', ['informations' => $informations, 'nomor' => $no]);
+        abort(403, 'Anda tidak memiliki cukup hak akses');
     }
 
     /**
@@ -33,7 +37,10 @@ class InformationController extends Controller
      */
     public function create()
     {
-        return view('informations.create');
+        if (Gate::allows('manage-informations')) {
+            return view('informations.create');
+        }
+        abort(403, 'Anda tidak memiliki cukup hak akses');
     }
 
     /**
@@ -44,18 +51,21 @@ class InformationController extends Controller
      */
     public function store(Request $request)
     {
-        \Validator::make($request->all(), [
-            'title' => 'required',
-            'desc' => 'required'
-        ])->validate();
+        if (Gate::allows('manage-informations')) {
+            \Validator::make($request->all(), [
+                'title' => 'required',
+                'desc' => 'required'
+            ])->validate();
 
-        $new_information = new \App\Information;
-        $new_information->title = $request->get('title');
-        $new_information->desc = $request->get('desc');
-        $new_information->date = Carbon::now();
+            $new_information = new \App\Information;
+            $new_information->title = $request->get('title');
+            $new_information->desc = $request->get('desc');
+            $new_information->date = Carbon::now();
 
-        $new_information->save();
-        return redirect()->route('informations.index')->with('success', 'Berita baru telah di tambah');
+            $new_information->save();
+            return redirect()->route('informations.index')->with('success', 'Berita baru telah di tambah');
+        }
+        abort(403, 'Anda tidak memiliki cukup hak akses');
     }
 
     /**
@@ -77,8 +87,11 @@ class InformationController extends Controller
      */
     public function edit($id)
     {
-        $information = \App\Information::findOrFail($id);
-        return view('informations.edit', ['information' => $information]);
+        if (Gate::allows('manage-informations')) {
+            $information = \App\Information::findOrFail($id);
+            return view('informations.edit', ['information' => $information]);
+        }
+        abort(403, 'Anda tidak memiliki cukup hak akses');
     }
 
     /**
@@ -90,18 +103,21 @@ class InformationController extends Controller
      */
     public function update(Request $request, $id)
     {
-        \Validator::make($request->all(), [
-            'title' => 'required',
-            'desc' => 'required'
-        ])->validate();
+        if (Gate::allows('manage-informations')) {
+            \Validator::make($request->all(), [
+                'title' => 'required',
+                'desc' => 'required'
+            ])->validate();
 
-        $information = \App\Information::findOrFail($id);
-        $information->title = $request->get('title');
-        $information->desc = $request->get('desc');
-        $information->date = Carbon::now();
+            $information = \App\Information::findOrFail($id);
+            $information->title = $request->get('title');
+            $information->desc = $request->get('desc');
+            $information->date = Carbon::now();
 
-        $information->save();
-        return redirect()->route('informations.index')->with('success', 'Berita berhasil di ubah');
+            $information->save();
+            return redirect()->route('informations.index')->with('success', 'Berita berhasil di ubah');
+        }
+        abort(403, 'Anda tidak memiliki cukup hak akses');
     }
 
     /**
@@ -112,10 +128,13 @@ class InformationController extends Controller
      */
     public function destroy($id)
     {
-        $information = \App\Information::findOrFail($id);
-        $information->delete();
+        if (Gate::allows('manage-informations')) {
+            $information = \App\Information::findOrFail($id);
+            $information->delete();
 
-        return redirect()->route('informations.index')->with('success', 'Berita berhasil di hapus');
+            return redirect()->route('informations.index')->with('success', 'Berita berhasil di hapus');
+        }
+        abort(403, 'Anda tidak memiliki cukup hak akses');
     }
 
     public function seeAllInformations()
